[UPDATE]完善考试判卷内容、更新用户提交查询、更新用户做题历史查询

+++ dcoj-web/src/main/java/com/dcoj/entity/exam/AnswerEntity.java
@@ -12,7 +12,7 @@ public class AnswerEntity {
 
     private Integer examProblemLocate;
 
-    private Byte problemType;
+    private Integer problemType;
 
     private Integer problemId;
